# apps/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Eduator Clean Backend API
  version: 0.4.0
  description: >-
    Clean backend API contract for local PostgreSQL + JWT + HTTP API keys (baseline **0.4.0**).
    Includes OpenRouter multi-model AI (migration `018_ai_provider_openrouter.sql`),
    pgvector-backed RAG (migration `015_pgvector_document_chunks.sql`), user HTTP API keys,
    usage analytics (`api_access_log` after migration `006_api_access_log.sql`),
    and first-party traffic exclusion via `X-Eduator-Client` (see docs).
servers:
  - url: http://localhost:4000
    description: Local backend
...
paths:
...
  /v1/ai/requests:
    post:
      tags: [AI]
      summary: Create generic AI request
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, payload]
              properties:
                type: { type: string }
                payload: { type: object, additionalProperties: true }
      responses:
        '201':
          description: Created
  /v1/ai/requests/{id}:
    get:
      tags: [AI]
      summary: Get generic AI request by ID
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Found
        '404':
          description: Not found
  /v1/ai/rag/retrieve:
    post:
      tags: [AI]
      summary: Retrieve relevant chunks from a document
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [documentId, query]
              properties:
                documentId:
                  type: string
                  format: uuid
                query:
                  type: string
                topK:
                  type: integer
                  default: 5
      responses:
        '200':
          description: Retrieved chunks
  /v1/ai/chat/assistants:
    get:
      tags: [AI]
      summary: List AI tutor assistants
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Assistant list
    post:
      tags: [AI]
      summary: Create AI tutor assistant
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                documentIds:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '201':
          description: Created
  /v1/ai/chat/assistants/{assistantId}:
    get:
      tags: [AI]
      summary: Get assistant by ID
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: assistantId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Assistant detail
    patch:
      tags: [AI]
      summary: Update assistant
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: assistantId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                documentIds:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '200':
          description: Updated
    delete:
      tags: [AI]
      summary: Delete assistant
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: assistantId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Deleted
  /v1/ai/chat/assistants/{assistantId}/conversations:
    get:
      tags: [AI]
      summary: List conversations for assistant
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: assistantId
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: externalUserId
          schema:
            type: string
      responses:
        '200':
          description: Conversation list
    post:
      tags: [AI]
      summary: Create conversation thread
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: assistantId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                externalUserId:
                  type: string
      responses:
        '201':
          description: Created
  /v1/ai/chat/conversations:
    post:
      tags: [AI]
      summary: Legacy — create assistant + first conversation
      deprecated: true
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                documentIds:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '201':
          description: Created
  /v1/ai/chat/conversations/{id}:
    get:
      tags: [AI]
      summary: Get chat conversation by ID
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Conversation detail
        '404':
          description: Not found
    patch:
      tags: [AI]
      summary: Update chat conversation
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
      responses:
        '200':
          description: Conversation updated
        '404':
          description: Not found
    delete:
      tags: [AI]
      summary: Delete chat conversation thread
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Conversation deleted
        '404':
          description: Not found
  /v1/ai/chat/conversations/{id}/messages:
    post:
      tags: [AI]
      summary: Send a chat message
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
                documentIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                shortAnswer:
                  type: boolean
      responses:
        '200':
          description: Assistant response
  /v1/ai/lessons/generate:
    post:
      tags: [AI]
      summary: Generate lesson from document
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [topic]
              properties:
                documentId: { type: string, format: uuid }
                documentIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                topic: { type: string }
                language: { type: string, default: en }
                gradeLevel: { type: string }
                objectives: { type: string }
                corePrompt: { type: string }
                options:
                  type: object
                  properties:
                    includeImages: { type: boolean }
                    includeAudio: { type: boolean }
                    includeTables: { type: boolean }
                    includeFigures: { type: boolean }
                    includeCharts: { type: boolean }
                    contentLength:
                      type: string
                      enum: [short, medium, full]
      responses:
        '201':
          description: Generated
  /v1/ai/exams/generate:
    post:
      tags: [AI]
      summary: Generate exam
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateExamRequest'
      responses:
        '201':
          description: Generated
  /v1/ai/exams/translate:
    post:
      tags: [AI]
      summary: Translate exam questions
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [questions, targetLanguage]
              properties:
                questions:
                  type: array
                  items:
                    type: object
                    additionalProperties: true
                targetLanguage:
                  type: string
      responses:
        '200':
          description: Translated
  /v1/ai/education-plans/generate:
    post:
      tags: [AI]
      summary: Generate education plan
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [documentId, name]
              properties:
                documentId: { type: string, format: uuid }
                name: { type: string }
                language:
                  type: string
                  description: Output language code (e.g. en, az). Same effect as outputLanguage.
                outputLanguage:
                  type: string
                  description: Alias of language for clients that use this field from the web UI.
                periodMonths: { type: integer, minimum: 1, maximum: 24 }
                sessionsPerWeek: { type: integer, minimum: 1, maximum: 14 }
                hoursPerSession: { type: integer, minimum: 1, maximum: 8 }
      responses:
        '201':
          description: Generated
  /v1/ai/translate:
    post:
      tags: [AI]
      summary: Translate text
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text, toLanguage]
              properties:
                text: { type: string }
                toLanguage: { type: string }
      responses:
        '200':
          description: Translated
  /v1/ai/tts:
    post:
      tags: [AI]
      summary: TTS placeholder
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string }
      responses:
        '200':
          description: Placeholder response
  /v1/ai/stt:
    post:
      tags: [AI]
      summary: STT placeholder
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string }
      responses:
        '200':
          description: Placeholder response
  /v1/ai/image/generate:
    post:
      tags: [AI]
      summary: Rewrite image prompt (phase-1)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt: { type: string }
      responses:
        '200':
          description: Rewritten prompt
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GenerateExamRequest:
      type: object
      properties:
        documentId:
          type: string
          format: uuid
        documentIds:
          type: array
          items:
            type: string
            format: uuid
        documentText:
          type: string
        title:
          type: string
        gradeLevel:
          type: string
        language:
          type: string
          default: en
        questionCount:
          type: integer
          minimum: 1
          maximum: 50
          default: 10
        topics:
          type: array
          items:
            type: string
        customPrompt:
          type: string
        questionTypes:
          type: array
          items:
            type: string
            enum: [multiple_choice, true_false, multiple_select, fill_blank]
        difficultyDistribution:
          type: object
          properties:
            easy: { type: integer, minimum: 0 }
            medium: { type: integer, minimum: 0 }
            hard: { type: integer, minimum: 0 }